+++ include/secrets.example.h
@@ -3,3 +3,4 @@
 #define SECRET_WIFI_SSID ""
 #define SECRET_WIFI_PASS ""
 #define SECRET_OWM_API_KEY ""
+#define SECRET_OTA_PASSWORD "change-me"